addition : circle draw function

+++ includes/ft_mlx.h
@@ -6,7 +6,7 @@
 /*   By: emaniez <emaniez@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/02/20 13:48:34 by emaniez           #+#    #+#             */
-/*   Updated: 2015/02/24 16:54:19 by emaniez          ###   ########.fr       */
+/*   Updated: 2015/02/24 17:29:46 by emaniez          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -86,6 +86,6 @@ void			mlx_draw_line(t_coord one, t_coord two, t_env *mlx, int color);
 
 void			mlx_draw_circle(t_coord *ctr, int rad, t_env *mlx, int color);
 
-int				mlx_coord_2_addr(int x, int y, t_env *mlx);
+int				mlx_coord_2_addr(size_t x, size_t y, t_env *mlx);
 
 #endif
